Fix subtle bug in CFGAS_GEFont::m_CharWidthMap.
It is currently a map onto int32_t. Originally, it was a map onto
uint16_t (implemented as CFX_WordDiscreteArray, which was a typedef
for CFX_DiscreteArrayTemplate<FX_WORD>, and FW_WORD was a typedef
for unsigned short). Some of the logic in GetCharWidth() relies
on storing a -1 into this map, but getting back 65535 when read,
which can't happen with the wider signed type.

Broken at https://pdfium-review.googlesource.com/c/pdfium/+/2953.
The result is logically correct, but we no longer avoid lookups
that could have been avoided.

We'll need to re-write this to something saner presently.

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/81450
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ xfa/fgas/font/cfgas_gefont.h
@@ -75,7 +75,7 @@ class CFGAS_GEFont final : public Retainable {
   RetainPtr<CPDF_Font> m_pPDFFont;  // Must come before |m_pFont|.
   MaybeOwned<CFX_Font> m_pFont;     // Must come before |m_pFontEncoding|.
   std::unique_ptr<CFX_UnicodeEncodingEx> m_pFontEncoding;
-  std::map<wchar_t, int32_t> m_CharWidthMap;
+  std::map<wchar_t, uint16_t> m_CharWidthMap;
   std::map<wchar_t, FX_RECT> m_BBoxMap;
   std::vector<RetainPtr<CFGAS_GEFont>> m_SubstFonts;
   std::map<wchar_t, RetainPtr<CFGAS_GEFont>> m_FontMapper;
